Creating and integrating new profiler module
The profiler can log simple events with a name, type, and timestamp.
It provides a mechanism to iterate over events with an arbitrary function.
This is used by the interpreter to print out all events with a routine it defines.

Events tracked: foreground thread start, periodic thread start, periodic thread end